Upstream tarball 20080405

+++ src/kademlia/kademlia/Kademlia.cpp
@@ -261,7 +261,7 @@ uint32 CKademlia::GetKademliaUsers(void)
 uint32 CKademlia::GetKademliaFiles(void)
 {
 	if( instance && instance->m_prefs ) {
-		return instance->m_prefs->GetKademliaFiles();
+		return (uint32)instance->m_prefs->GetKademliaFiles();
 	}
 	return 0;
 }
